requirements.txt
backports.zoneinfo;python_version<"3.9"
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
decorator
gitdb
GitPython
idna
importlib-metadata
importlib-resources
Jinja2
jsonschema
jsonschema-specifications
lxml
markdown-it-py
MarkupSafe
mdurl
numpy
packaging
pandas
pdf2image
Pillow
pkgutil_resolve_name
protobuf
pyarrow
pydeck
Pygments
Pympler
python-dateutil
python-docx
pytz
pytz-deprecation-shim
referencing
requests
rich
rpds-py
six
smmap
streamlit
tenacity
toml
toolz
tornado
typing_extensions
tzdata
tzlocal
urllib3
validators
zipp
